pyproject.toml
[project]
name = "dataviz-backend"
version = "0.1.0"
description = "LLM-powered data visualization backend with multi-agent system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pandas>=2.2.0",
    "plotly>=5.18.0",
    "google-generativeai>=0.3.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "kaleido>=0.2.1",  # Pour export PNG
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dataviz_backend"]